+++ e2e/internal/devnet/smartcontract_geolocation.go
@@ -81,7 +81,7 @@ func (dn *Devnet) InitGeolocationProgramConfigIfNotInitialized(ctx context.Conte
 	}, docker.NoPrintOnError())
 	if err != nil {
 		outputStr := strings.ToLower(string(output))
-		if strings.Contains(outputStr, "already") || strings.Contains(outputStr, "already in use") {
+		if strings.Contains(outputStr, "already") || strings.Contains(outputStr, "already in use") || strings.Contains(outputStr, "uninitialized account") {
 			dn.log.Debug("--> Geolocation program config is already initialized")
 			return false, nil
 		}